Add coffee bean detail page'

+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/farhanazyumardhiazmi/coffeepedia/data/dao/CoffeeBeanDao.java
@@ -22,5 +22,5 @@ public interface CoffeeBeanDao {
 	LiveData<List<CoffeeBean>> loadAllCoffeeBeans();
 
 	@Query("SELECT * FROM coffee_beans WHERE coffeeBeanId = :coffeeBeanId")
-	LiveData<CoffeeBean> getCoffeeBeanById(int coffeeBeanId);
+	LiveData<CoffeeBean> getCoffeeBeanById(long coffeeBeanId);
 }
