+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/farhanazyumardhiazmi/coffeepedia/data/repository/CoffeeBeanRepository.java
@@ -33,7 +33,7 @@ public class CoffeeBeanRepository {
 		return mObservableCoffeeBeans;
 	}
 
-	public LiveData<CoffeeBean> getCoffeeBeanById(final int coffeeBeanId) {
+	public LiveData<CoffeeBean> getCoffeeBeanById(final long coffeeBeanId) {
 		return mDatabase.coffeeBeanDao().getCoffeeBeanById(coffeeBeanId);
 	}
